backend/requirements.txt
# Backend Dependencies - Förder-Finder Grundschule

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6

# Database
cx_Oracle==8.3.0
python-dotenv==1.0.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0

# Scraping
scrapy==2.11.0
scrapy-user-agents==0.1.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# RAG & AI
langchain==0.1.0
chromadb==0.4.18
sentence-transformers==2.2.2
httpx==0.25.2
openai==1.5.0

# Vector Embeddings
torch==2.1.1
transformers==4.35.2

# OCI SDK
oci==2.115.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2
pytz==2023.3

# Logging & Monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx-mock==0.11.0
faker==20.1.0
